fix: 카드 중복발급 되지 않도록 수정
